Added GPIO code plus general structure of loop
Added the GPIO code and addresses to control LEDs and buttons in the system. Also added the general structure for state behavior and transitions inside of the main loop(). Added comments of what should be done and where in order to have expected behavior. Now those behaviors just need to be implemented.