Code Using Libraries
This code was developed with both partners working on it in person together as a collaborative effort.
This code includes most of the functionality of the swamp cooler but uses libraries. This will be changed later and is only a temporary file as it includes libraries that are not allowed in the code.
The working parts include but are not limited to state chagning being refelcted by the LCD and LEDs, the monitoring of the correct sensors based on each state, inputs correctly working, and the motor enabling durring the correct states.
Some functionality is missing from the code and comments at the top of the file note some of the required changes. All pins are based on the circuit diagram.
This code has been tested on the prototype of the system developed and works to some extent (some errors are listed in comments at the top of the file.